Use Maven magic to externalize some properties
And allow them to be read from files not under version control and templated into a service that's accessible through Angular dependency injection. The Java code we can take care of through Spring DI. This allows us to use a single set of settings for both client and server while keeping them confidential.